requirements.txt
streamlit>=1.31.0
langchain-google-genai>=0.0.11
python-dotenv>=1.0.0
selenium>=4.18.1
pydantic>=2.6.1 